refactor: 마이페이지 탈퇴 플로우 재구성 #21
